# Autoria-Final/AutoriaFinal-Frontend/unified-app/apidoc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Autoria Auction API",
    "description": "API for auction management system",
    "version": "v1"
  },
  "paths": {
...
    "/api/Auction/{id}": {
      "get": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/AuctionCar/auction/{auctionId}": {
      "get": {
        "tags": [
          "AuctionCar"
        ],
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionCarGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionCarGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionCarGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Car": {
      "get": {
        "tags": [
          "Car"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Car"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Make",
                  "Model",
                  "Vin",
                  "Year"
                ],
                "type": "object",
                "properties": {
                  "Vin": {
                    "maxLength": 17,
                    "minLength": 0,
                    "type": "string"
                  },
                  "Year": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Make": {
                    "maxLength": 100,
                    "minLength": 0,
                    "type": "string"
                  },
                  "Model": {
                    "maxLength": 100,
                    "minLength": 0,
                    "type": "string"
                  },
                  "BodyStyle": {
                    "type": "string"
                  },
                  "Color": {
                    "type": "string"
                  },
                  "LocationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "Image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "ImagePath": {
                    "type": "string"
                  },
                  "OwnerId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Vin": {
                  "style": "form"
                },
                "Year": {
                  "style": "form"
                },
                "Make": {
                  "style": "form"
                },
                "Model": {
                  "style": "form"
                },
                "BodyStyle": {
                  "style": "form"
                },
                "Color": {
                  "style": "form"
                },
                "LocationId": {
                  "style": "form"
                },
                "Image": {
                  "style": "form"
                },
                "ImagePath": {
                  "style": "form"
                },
                "OwnerId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Car/{id}": {
      "get": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CarUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CarUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Car/vin/{vin}": {
      "get": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "vin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Car/{id}/photo": {
      "post": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Car/{id}/video": {
      "post": {
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuctionCarGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "auctionId": {
            "type": "string",
            "format": "uuid"
          },
          "carId": {
            "type": "string",
            "format": "uuid"
          },
          "lotNumber": {
            "type": "string",
            "nullable": true
          },
          "currentPrice": {
            "type": "number",
            "format": "double"
          },
          "minPreBid": {
            "type": "number",
            "format": "double"
          },
          "winnerStatus": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "bidCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastBidTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isReserveMet": {
            "type": "boolean"
          },
          "reservePrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "carMake": {
            "type": "string",
            "nullable": true
          },
          "carModel": {
            "type": "string",
            "nullable": true
          },
          "carYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "carImage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuctionDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "startTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "minBidIncrement": {
            "type": "number",
            "format": "double"
          },
          "startPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "timerSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "currentCarLotNumber": {
            "type": "string",
            "nullable": true
          },
          "isLive": {
            "type": "boolean"
          },
          "extendedCount": {
            "type": "integer",
            "format": "int32"
          },
          "maxCarDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "currentCarStartTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "totalCarsCount": {
            "type": "integer",
            "format": "int32"
          },
          "carsWithPreBidsCount": {
            "type": "integer",
            "format": "int32"
          },
          "soldCarsCount": {
            "type": "integer",
            "format": "int32"
          },
          "unsoldCarsCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalSalesAmount": {
            "type": "number",
            "format": "double"
          },
          "auctionCars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuctionCarGetDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuctionUpdateDto": {
        "type": "object",
        "additionalProperties": false
      },
...
      "CarCondition": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "CarUpdateDto": {
        "type": "object",
        "properties": {
          "vin": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "bodyStyle": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "odometer": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "odometerUnit": {
            "type": "string",
            "nullable": true
          },
          "fuel": {
            "$ref": "#/components/schemas/FuelType"
          },
          "transmission": {
            "$ref": "#/components/schemas/Transmission"
          },
          "driveTrain": {
            "$ref": "#/components/schemas/DriveTrain"
          },
          "condition": {
            "$ref": "#/components/schemas/CarCondition"
          },
          "hasKeys": {
            "type": "boolean"
          },
          "primaryDamage": {
            "$ref": "#/components/schemas/DamageType"
          },
          "secondaryDamage": {
            "$ref": "#/components/schemas/DamageType"
          },
          "titleType": {
            "$ref": "#/components/schemas/TitleType"
          },
          "titleState": {
            "type": "string",
            "nullable": true
          },
          "estimatedRetailValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "imageCar": {
            "type": "string",
            "format": "binary",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DamageType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          13
        ],
        "type": "integer",
        "format": "int32"
      },
      "DriveTrain": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "FuelType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "TitleType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Transmission": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}